+++ 01 Main Files/402nd_Armory/CfgAddons.hpp
@@ -1,0 +1,10 @@
+class CfgAddons
+{
+	class PreloadAddons
+	{
+		class 402nd_Armory
+		{
+			list[]={"402nd_Armory"};
+		};
+	};
+};
