+++ 01 Main Files/402nd_Armory/CfgPatches.hpp
@@ -1,0 +1,18 @@
+class CfgPatches
+{
+	class 402nd_Armory
+	{
+		author="Azzamean, Vespade & BoomMicGuy"
+		name="402nd Armory"
+		url="https://discord.gg/fa7Cguk";
+		// Added Items
+		units[]={};
+		weapons[]={};
+		magazines[]={};
+		ammo[]={};
+		// Requirements
+		requiredVersion=0.1;
+		requiredAddons[]=
+		{"V_FZ_Armor","V_FZ_Core","OPTRE_UNSC_Units_Army"};
+	};
+};
